+++ utils/database/migrations.py
@@ -23,6 +23,8 @@ SOFTWARE.
 """
 
 from utils.logging_formatter import bot_logger
+from typing import Type, List
+from types import TracebackType
 import aiosqlite
 import aiofiles
 import asyncio
@@ -30,24 +32,77 @@ import os
 
 
 class MigrationVersionMismatch(Exception):
+    """
+    Error raised when the next migration is not sequential relative to the current version.
+    """
+
     pass
 
+
 class Migration:
+    """
+    A class that holding information related to a SQL migration file.
+
+    Attributes:
+        script (str): The (raw) SQL script to be executed.
+        _version (str): The version associated with this script.
+        name (str): The name of this migration.
+    """
 
     def __init__(self, script: str, version: str, name: str) -> None:
+        """
+        The constructor for the Migration class.
+
+        Parameters:
+            script (str): The (raw) SQL script to be executed.
+            version (str): The version associated with this script.
+            name (str): The name of this migration.
+        """
+
         self.script = script
         self._version = version
         self.name = name
 
     def __repr__(self):
+        """
+        Produces a printable representation of this class.
+
+        Parameters:
+            None.
+
+        Returns:
+            (str).
+        """
+
         return f'{self._version}-{self.name}'
 
     @property
     def version(self) -> int:
+        """
+        Provides integer access to this migration's version.
+
+        Parameters:
+            None.
+
+        Returns:
+            (int).
+        """
+
         return int(self._version)
 
     @classmethod
-    async def from_file(cls, path: str, file: str):
+    async def from_file(cls, path: str, file: str) -> 'Migration':
+        """
+        Creates a Migration from the respective filepath.
+
+        Parameters:
+            path (str): The 'migrations' directory filepath.
+            file (str): The name of the migration file.
+
+        Returns:
+            (Migration).
+        """
+
         version, name = file.split('-', 1)
 
         async with aiofiles.open(os.path.join(path, file), mode='r') as f:
@@ -57,18 +112,62 @@ class Migration:
 
 
 class Migrator:
+    """
+    A class that manages SQL migration storage, preparation, and execution.
 
-    def __init__(self, database):
+    Attributes:
+        database (str): The name of the database file.
+        version (int): The current database version.
+        _migrations (List[Migration]): A list of all successfully-parsed migration files.
+    """
+
+    def __init__(self, database: str):
+        """
+        The constructor for the Migrator class.
+
+        Parameters:
+            database (str): The name of the database file.
+
+        Returns:
+            None.
+        """
+
         self.database = database
         self.version = 0
         self._migrations = []
 
-    async def __aenter__(self):
+    async def __aenter__(self) -> 'Migrator':
+        """
+        Asynchronous entry point logic.
+
+        Prepares the internal list of Migrations when executed.
+
+        Parameters:
+            None.
+
+        Returns:
+            (Migrator).
+        """
+
         await self._prepare_migrations()
 
         return self
 
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
+    async def __aexit__(self, exc_type: Type, exc_val: Exception, exc_tb: TracebackType) -> None:
+        """
+        Asynchronous exit point logic.
+
+        Prepares the internal list of Migrations when executed.
+
+        Parameters:
+            exc_type (Type): The type of exception raised.
+            exc_val (Exception): The exception raised.
+            exc_tb (TracebackType): The traceback of the exception raised.
+
+        Returns:
+            None.
+        """
+
         if exc_type:
             exc_val = exc_val or 'N/A'
             bot_logger.error(f'Encountered {exc_type}: {exc_val} while applying migrations.')
@@ -76,10 +175,35 @@ class Migrator:
             bot_logger.error(f'Traceback: {exc_tb}')
 
     @property
-    def _available_migrations(self) -> [Migration]:
+    def _available_migrations(self) -> List[Migration]:
+        """
+        Returns a filtered list of available migrations.
+        An available migration is a migration with a version greater than the current database version.
+
+        Parameters:
+            None.
+
+        Returns:
+            (List[Migration]).
+        """
+
         return sorted([x for x in self._migrations if x.version > self.version], key=lambda x: x.version)
 
     async def apply_migrations(self) -> None:
+        """
+        Applies all available migrations to the database.
+
+        There's an implicit first migration of creating (if necessary) the database and fetching the current version.
+        Afterwards, if there are still available migrations, a 10 second 'abort' period starts.
+        If the script is still active after 10 seconds, migrations are applied atomically.
+
+        Parameters:
+            None.
+
+        Returns:
+            None.
+        """
+
         await self._create_database()
 
         migration_count = len(self._available_migrations)
@@ -105,13 +229,35 @@ class Migrator:
         bot_logger.info(f'Successfully applied all migrations.')
 
     async def _create_database(self) -> None:
+        """
+        Connects to (and implicitly creates if necessary) and fetches the current database version.
+
+        Parameters:
+            None.
+
+        Returns:
+            None.
+        """
+
         async with aiosqlite.connect(self.database) as connection:
             async with connection.execute('PRAGMA user_version') as cursor:
                 self.version = (await cursor.fetchone())[0]
 
     async def _migrate(self, migration: Migration) -> None:
+        """
+        Applies a given migration.
+
+        If the current migration is not strictly sequential relative to the current version, the migration fails.
+
+        Parameters:
+            migration (Migration): The migration to apply.
+
+        Returns:
+            None.
+        """
+
         if self.version != migration.version - 1:
-            raise MigrationVersionMismatch(f'{migration}')
+            raise MigrationVersionMismatch(f'{migration} [{self.version} -> {migration.version}]')
 
         async with aiosqlite.connect(self.database) as connection:
             await connection.executescript(migration.script)
@@ -119,6 +265,16 @@ class Migrator:
                 self.version = (await cursor.fetchone())[0]
 
     async def _prepare_migrations(self) -> None:
+        """
+        Generates a list of Migrations from the bot's 'migrations' directory.
+
+        Parameters:
+            None.
+
+        Returns:
+            None.
+        """
+
         path = os.path.join(os.getcwd(), 'migrations')
         files = [f for f in os.listdir(path) if os.path.isfile(os.path.join(path, f))]